+++ doc/CHANGELOG.md
@@ -77,6 +77,7 @@ The full set of tests is only run for tagged releases.
   - move `using namespace std` after includes (clang warning)
   - some selected clang -Weverything aspects
   - now -Wdocumentation clean (some wrong doxygen trailing comments)
+- rw11/shell.tcl: add workaround for tclreadline and `after` interference
 
 ### Bug Fixes
 - RtclArgs.hpp: BUGFIX: get *_min limits correct (gcc -Wpedantic)